Rename MASynergy, add a little decoration to clarify its work.
