+++ src/Game.ts
@@ -55,7 +55,7 @@ import { OrOptions } from "./inputs/OrOptions";
 import { SelectOption } from "./inputs/SelectOption";
 import { LogHelper } from "./components/LogHelper";
 import { ColonyName } from "./colonies/ColonyName";
-import { getRandomMilestonesAndAwards } from "./MASynergy";
+import { getRandomMilestonesAndAwards } from "./MilestoneAwardSelecter";
 import { CardType } from "./cards/CardType";
 import { ColonyModel } from "./models/ColonyModel";
 import { LogBuilder } from "./LogBuilder";
@@ -393,9 +393,9 @@ export class Game implements ILoadable<SerializedGame, Game> {
     }
 
     public setRandomMilestonesAndAwards(hasVenus: boolean, requiredQty: number) {
-      const MA_Info = getRandomMilestonesAndAwards(hasVenus, requiredQty);
-      this.milestones.push(...MA_Info.milestones);
-      this.awards.push(...MA_Info.awards);
+      const drawnMilestonesAndAwards = getRandomMilestonesAndAwards(hasVenus, requiredQty);
+      this.milestones.push(...drawnMilestonesAndAwards.milestones);
+      this.awards.push(...drawnMilestonesAndAwards.awards);
     }
 
     // Add Venus Next board colonies and milestone / award
