+++ tests/MilestoneAwardSelecter.spec.ts
@@ -1,7 +1,7 @@
 import { expect } from "chai";
-import { buildSynergies, SYNERGIES } from "../src/MASynergy";
+import { buildSynergies, SYNERGIES } from "../src/MilestoneAwardSelecter";
 
-describe("MASynergy", function () {
+describe("MilestoneAwardSelecter", function () {
     it("verify that hand-built synergies is the same as the predefined array", function () {
         expect(buildSynergies()).is.eql(SYNERGIES);
     });
